refactor(ccxt): migrate modules into dedicated package
